Refactor request and response into root
